add epoch test, refactor
